Cast uint32 to float.

+++ src/server/game/Entities/Unit/Unit.cpp
@@ -12528,7 +12528,7 @@ int32 Unit::HealBySpell(Unit* victim, SpellInfo const* spellInfo, uint32 addHeal
     uint32 absorb = 0;
     int32 gain = 0;
 
-    sScriptMgr->ModifyHealReceived(this, victim, addHealth);
+    sScriptMgr->ModifyHealReceived(this, victim, static_cast<float>(addHealth));
 
     // calculate heal absorb and reduce healing
     CalcHealAbsorb(victim, spellInfo, addHealth, absorb);
